tradehut_search/requirements.txt
# Search — Django side
# pip install -r requirements.txt

# Elasticsearch
elasticsearch-dsl>=8.13,<9
django-elasticsearch-dsl>=7.3,<8

# Image search (pgvector + CLIP)
pgvector>=0.3.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.38.0
Pillow>=10.0.0

# HTTP client for calling the embedding service
httpx>=0.27.0

# Async tasks
celery>=5.3.0
redis>=5.0.0
